app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.8.0
factory-boy==3.2.0
Faker==8.14.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==3.2.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
unicodecsv==0.14.1
uritemplate==3.0.1
